# index.ts
import { Hono } from 'hono'
import { handleRequest } from './src/proxy.js'
import { startWatcher, getRecentChanges } from './src/watcher.js'

const app = new Hono()

// Auth middleware (applied to /proxy/*)
const authMiddleware = async (c, next) => {
  const auth = c.req.header('Authorization')
  if (auth !== `Bearer ${process.env.API_KEY}`) {
    return c.json({ error: 'Unauthorized' }, 401)
  }
  await next()
}

// Apply auth to all /proxy routes
app.use('/proxy/*', authMiddleware)

// Serve OpenAPI spec
app.get('/openapi.json', async (c) => {
  const yaml = await Bun.file('./openapi.yaml').text();
  const spec = JSON.parse(yaml);
  return c.json(spec, 200, { 'Content-Type': 'application/json' });
})

app.get('/', (c) => c.text('Spar API Proxy'))

app.get('/proxy/health', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const result = await handleRequest('health', {}, base)
  return c.json(result)
})

app.post('/proxy/files/read', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('files/read', body, base)
  return c.json(result)
})

app.post('/proxy/files/write', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('files/write', body, base)
  return c.json(result)
})

app.delete('/proxy/files/:path*', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const pathParam = c.req.param('path') || ''
  const result = await handleRequest('files/delete', { path: pathParam }, base)
  return c.json(result)
})

app.get('/proxy/dirs/:path*', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const pathParam = c.req.param('path') || '.'
  const query = c.req.query()
  const result = await handleRequest('dirs/list', { path: pathParam, options: { recursive: query.recursive === 'true', includeFiles: query.includeFiles !== 'false' } }, base)
  return c.json(result)
})

app.post('/proxy/search/glob', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const ignore = body.ignore || []
  const result = await handleRequest('search/glob', { pattern: body.pattern, path: body.path || base, ignore }, base)
  return c.json(result)
})

app.post('/proxy/search/grep', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const ignore = body.ignore || []
  const result = await handleRequest('search/grep', { pattern: body.pattern, path: body.path || base, include: body.include, ignore }, base)
  return c.json(result)
})

app.post('/proxy/exec/start', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('exec/start', { ...body, infinite: body.infinite }, base)
  return c.json(result)
})

app.get('/proxy/exec/:jobId/status', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const jobId = c.req.param('jobId')
  const result = await handleRequest('exec/status', { jobId }, base)
  return c.json(result)
})

app.post('/proxy/exec/:jobId/stop', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const jobId = c.req.param('jobId')
  const result = await handleRequest('exec/stop', { jobId }, base)
  return c.json(result)
})

app.post('/proxy/exec/bash', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('exec/bash', body, base)
  return c.json(result)
})

app.get('/proxy/git/status', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const result = await handleRequest('git/status', {}, base)
  return c.json(result)
})

app.post('/proxy/git/diff', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('git/diff', body, base)
  return c.json(result)
})

app.post('/proxy/git/commit', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('git/commit', body, base)
  return c.json(result)
})

app.get('/proxy/git/log', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const query = c.req.query()
  const result = await handleRequest('git/log', { limit: parseInt(query.limit || '10') }, base)
  return c.json(result)
})

app.post('/proxy/lsp/diagnostics', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('lsp/diagnostics', body, base)
  return c.json(result)
})

app.post('/proxy/lsp/definition', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('lsp/definition', body, base)
  return c.json(result)
})

app.post('/proxy/lsp/references', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const body = await c.req.json()
  const result = await handleRequest('lsp/references', body, base)
  return c.json(result)
})

app.get('/proxy/events', async (c) => {
  const base = process.env.WORKSPACE_PATH || '/workspace'
  const query = c.req.query()
  const since = parseInt(query.since || '0')
  const changes = getRecentChanges(since)
  return c.json({ changes })
})

const base = process.env.WORKSPACE_PATH || '/workspace'
startWatcher(base)

const port = parseInt(process.env.PORT || '3000')
Bun.serve({
  fetch: app.fetch,
  port,
})

console.log(`Spar API Proxy running on http://localhost:${port}`)

export default app