# openapi.yaml
openapi: 3.0.0
info:
  title: Spar API Proxy
  version: 1.0.0
  description: Secure API proxy for agentic coding tools to interact with mounted /workspace codebase.
servers:
  - url: http://localhost:3000
    description: Local server
paths:
  /proxy/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  workspace:
                    type: object
                    properties:
                      exists:
                        type: boolean
                      writable:
                        type: boolean
  /proxy/files/read:
    post:
      summary: Read one or multiple files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                paths:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Files read
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        content:
                          type: string
                        error:
                          type: string
  /proxy/files/write:
    post:
      summary: Write/overwrite file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /proxy/files/{path}:
    delete:
      summary: Delete file
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /proxy/dirs/{path}:
    get:
      summary: List directory
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
        - name: recursive
          in: query
          schema:
            type: boolean
            default: false
        - name: includeFiles
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Directory listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: string
                  dirs:
                    type: array
                    items:
                      type: string
  /proxy/search/glob:
    post:
      summary: Glob file search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                path:
                  type: string
                ignore:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Matches
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      type: string
  /proxy/search/grep:
    post:
      summary: Grep content search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                path:
                  type: string
                include:
                  type: string
                ignore:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Matches
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      type: object
                      properties:
                        file:
                          type: string
                        lines:
                          type: array
                          items:
                            type: object
                            properties:
                              line:
                                type: number
                              content:
                                type: string
  /proxy/exec/start:
    post:
      summary: Start background exec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                timeout:
                  type: number
      responses:
        '200':
          description: Job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
  /proxy/exec/{jobId}/status:
    get:
      summary: Get exec status and output
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  stdout:
                    type: string
                  stderr:
                    type: string
                  exitCode:
                    type: number
                  duration:
                    type: number
                  error:
                    type: string
  /proxy/exec/bash:
    post:
      summary: Blocking exec (legacy)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                timeout:
                  type: number
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                type: object
                properties:
                  stdout:
                    type: string
                  stderr:
                    type: string
                  exitCode:
                    type: number
                  duration:
                    type: number
  /proxy/git/status:
    get:
      summary: Git status
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  staged:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        status:
                          type: string
                  unstaged:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        status:
                          type: string
                  untracked:
                    type: array
                    items:
                      type: string
  /proxy/git/diff:
    post:
      summary: Git diff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
      responses:
        '200':
          description: Diff
          content:
            application/json:
              schema:
                type: object
                properties:
                  diff:
                    type: string
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        file:
                          type: string
                        changes:
                          type: string
  /proxy/git/commit:
    post:
      summary: Git commit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                all:
                  type: boolean
      responses:
        '200':
          description: Commit result
          content:
            application/json:
              schema:
                type: object
                properties:
                  commitHash:
                    type: string
                  success:
                    type: boolean
  /proxy/git/log:
    get:
      summary: Git log
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        '200':
          description: Log
          content:
            application/json:
              schema:
                type: object
                properties:
                  commits:
                    type: array
                    items:
                      type: object
                      properties:
                        hash:
                          type: string
                        message:
                          type: string
                        author:
                          type: string
                        date:
                          type: string
  /proxy/lsp/diagnostics:
    post:
      summary: LSP diagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: Diagnostics
          content:
            application/json:
              schema:
                type: object
                properties:
                  diagnostics:
                    type: array
                    items:
                      type: object
                      properties:
                        severity:
                          type: string
                        message:
                          type: string
                        range:
                          type: object
                          properties:
                            start:
                              type: object
                              properties:
                                line:
                                  type: number
                                character:
                                  type: number
                            end:
                              type: object
                              properties:
                                line:
                                  type: number
                                character:
                                  type: number
                        code:
                          type: string
  /proxy/lsp/definition:
    post:
      summary: LSP definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                line:
                  type: number
                character:
                  type: number
                content:
                  type: string
      responses:
        '200':
          description: Locations
          content:
            application/json:
              schema:
                type: object
                properties:
                  locations:
                    type: array
                    items:
                      type: object
                      properties:
                        uri:
                          type: string
                        range:
                          type: object
                          properties:
                            start:
                              type: object
                              properties:
                                line:
                                  type: number
                                character:
                                  type: number
                            end:
                              type: object
                              properties:
                                line:
                                  type: number
                                character:
                                  type: number
  /proxy/lsp/references:
    post:
      summary: LSP references
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                line:
                  type: number
                character:
                  type: number
                content:
                  type: string
      responses:
        '200':
          description: References
          content:
            application/json:
              schema:
                type: object
                properties:
                  references:
                    type: array
                    items:
                      type: object
                      properties:
                        uri:
                          type: string
                        range:
                          type: object
                          properties:
                            start:
                              type: object
                              properties:
                                line:
                                  type: number
                                character:
                                  type: number
                            end:
                              type: object
                              properties:
                                line:
                                  type: number
                                character:
                                  type: number
  /proxy/events:
    get:
      summary: Recent file changes
      parameters:
        - name: since
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Changes
          content:
            application/json:
              schema:
                type: object
                properties:
                  changes:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        type:
                          type: string
                        timestamp:
                          type: number
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []